refactor: conflict로 인한 오류 해결
