+++ backend/pokerogue/src/test/java/com/pokerogue/helper/move/converter/MoveCategoryConverterTest.java
@@ -2,7 +2,7 @@ package com.pokerogue.helper.move.converter;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.pokerogue.helper.battle.MoveCategory;
+import com.pokerogue.helper.move.data.MoveCategory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
