+++ backend/pokerogue/src/test/java/com/pokerogue/helper/move/repository/MoveRepositoryTest.java
@@ -4,9 +4,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 import com.pokerogue.environment.repository.MongoRepositoryTest;
-import com.pokerogue.helper.battle.MoveCategory;
 import com.pokerogue.helper.move.MoveRepository;
 import com.pokerogue.helper.move.data.Move;
+import com.pokerogue.helper.move.data.MoveCategory;
 import com.pokerogue.helper.move.data.MoveFlag;
 import com.pokerogue.helper.move.data.MoveTarget;
 import com.pokerogue.helper.type.data.Type;
